@@ -12,9 +12,9 @@ import 'liquid_progress_indicator.dart';
 ///
 class SimpleAppUpgradeWidget extends StatefulWidget {
   const SimpleAppUpgradeWidget(
-      {@required this.title,
+      {required this.title,
       this.titleStyle,
-      @required this.contents,
+      required this.contents,
       this.contentStyle,
       this.cancelText,
       this.cancelTextStyle,
@@ -26,7 +26,7 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
       this.borderRadius = 10,
       this.downloadUrl,
       this.force = false,
-      this.iosAppId,
+      required this.iosAppId,
       this.appMarketInfo,
       this.onCancel,
       this.onOk,
@@ -41,7 +41,7 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   ///
   /// 标题样式
   ///
-  final TextStyle titleStyle;
+  final TextStyle? titleStyle;
 
   ///
   /// 升级提示内容
@@ -51,47 +51,47 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   ///
   /// 提示内容样式
   ///
-  final TextStyle contentStyle;
+  final TextStyle? contentStyle;
 
   ///
   /// 下载进度条
   ///
-  final Widget progressBar;
+  final Widget? progressBar;
 
   ///
   /// 进度条颜色
   ///
-  final Color progressBarColor;
+  final Color? progressBarColor;
 
   ///
   /// 确认控件
   ///
-  final String okText;
+  final String? okText;
 
   ///
   /// 确认控件样式
   ///
-  final TextStyle okTextStyle;
+  final TextStyle? okTextStyle;
 
   ///
   /// 确认控件背景颜色,2种颜色左到右线性渐变
   ///
-  final List<Color> okBackgroundColors;
+  final List<Color>? okBackgroundColors;
 
   ///
   /// 取消控件
   ///
-  final String cancelText;
+  final String? cancelText;
 
   ///
   /// 取消控件样式
   ///
-  final TextStyle cancelTextStyle;
+  final TextStyle? cancelTextStyle;
 
   ///
   /// app安装包下载url,没有下载跳转到应用宝等渠道更新
   ///
-  final String downloadUrl;
+  final String? downloadUrl;
 
   ///
   /// 圆角半径
@@ -112,12 +112,12 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   /// 指定跳转的应用市场，
   /// 如果不指定将会弹出提示框，让用户选择哪一个应用市场。
   ///
-  final AppMarketInfo appMarketInfo;
+  final AppMarketInfo? appMarketInfo;
 
-  final VoidCallback onCancel;
-  final VoidCallback onOk;
-  final DownloadProgressCallback downloadProgress;
-  final DownloadStatusChangeCallback downloadStatusChange;
+  final VoidCallback? onCancel;
+  final VoidCallback? onOk;
+  final DownloadProgressCallback? downloadProgress;
+  final DownloadStatusChangeCallback? downloadStatusChange;
 
   @override
   State<StatefulWidget> createState() => _SimpleAppUpgradeWidget();
@@ -154,16 +154,18 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
   ///
   Widget _buildInfoWidget(BuildContext context) {
     return Container(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          //标题
-          _buildTitle(),
-          //更新信息
-          _buildAppInfo(),
-          //操作按钮
-          _buildAction()
-        ],
+      child: SingleChildScrollView(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            //标题
+            _buildTitle(),
+            //更新信息
+            _buildAppInfo(),
+            //操作按钮
+            _buildAction()
+          ],
+        ),
       ),
     );
   }
@@ -174,7 +176,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
   _buildTitle() {
     return Padding(
         padding: EdgeInsets.only(top: 20, bottom: 30),
-        child: Text(widget.title ?? '',
+        child: Text(widget.title,
             style: widget.titleStyle ?? TextStyle(fontSize: 22)));
   }
 
@@ -184,7 +186,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
   _buildAppInfo() {
     return Container(
         padding: EdgeInsets.only(left: 15, right: 15, bottom: 30),
-        height: 200,
+        constraints: BoxConstraints(maxHeight: 150, minHeight: 120),
         child: ListView(
           children: widget.contents.map((f) {
             return Text(
@@ -257,8 +259,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
           bottomLeft: Radius.circular(widget.borderRadius));
     }
     var _okBackgroundColors = widget.okBackgroundColors;
-    if (widget.okBackgroundColors == null ||
-        widget.okBackgroundColors.length != 2) {
+    if (_okBackgroundColors == null || _okBackgroundColors.length != 2) {
       _okBackgroundColors = [
         Theme.of(context).primaryColor,
         Theme.of(context).primaryColor
@@ -297,6 +298,8 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
           valueColor: AlwaysStoppedAnimation(widget.progressBarColor ??
               Theme.of(context).primaryColor.withOpacity(0.4)),
           borderRadius: widget.borderRadius,
+          borderColor: Colors.transparent,
+          borderWidth: 0,
         );
   }
 
@@ -310,13 +313,13 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
       FlutterUpgrade.toAppStore(widget.iosAppId);
       return;
     }
-    if (widget.downloadUrl == null || widget.downloadUrl.isEmpty) {
+    if (widget.downloadUrl == null || widget.downloadUrl!.isEmpty) {
       //没有下载地址，跳转到第三方渠道更新，原生实现
       FlutterUpgrade.toMarket(appMarketInfo: widget.appMarketInfo);
       return;
     }
     String path = await FlutterUpgrade.apkDownloadPath;
-    _downloadApk(widget.downloadUrl, '$path/$_downloadApkName');
+    _downloadApk(widget.downloadUrl!, '$path/$_downloadApkName');
   }
 
   ///
@@ -351,7 +354,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
     } catch (e) {
       print('$e');
       _downloadProgress = 0;
-      _updateDownloadStatus(DownloadStatus.error,error: e);
+      _updateDownloadStatus(DownloadStatus.error, error: e);
     }
   }
 
