@@ -8,10 +8,10 @@ class Wave extends StatefulWidget {
   final Axis direction;
 
   const Wave({
-    Key key,
-    @required this.value,
-    @required this.color,
-    @required this.direction,
+    Key? key,
+    required this.value,
+    required this.color,
+    required this.direction,
   }) : super(key: key);
 
   @override
@@ -19,7 +19,7 @@ class Wave extends StatefulWidget {
 }
 
 class _WaveState extends State<Wave> with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
+  late AnimationController _animationController;
 
   @override
   void initState() {
@@ -65,9 +65,9 @@ class _WaveClipper extends CustomClipper<Path> {
   final Axis direction;
 
   _WaveClipper({
-    @required this.animationValue,
-    @required this.value,
-    @required this.direction,
+    required this.animationValue,
+    required this.value,
+    required this.direction,
   });
 
   @override
@@ -94,7 +94,7 @@ class _WaveClipper extends CustomClipper<Path> {
     for (int i = -2; i <= size.height.toInt() + 2; i++) {
       final waveHeight = (size.width / 20);
       final dx = math.sin((animationValue * 360 - i) % 360 * (math.pi / 180)) *
-          waveHeight +
+              waveHeight +
           (size.width * value);
       waveList.add(Offset(dx, i.toDouble()));
     }
@@ -106,7 +106,7 @@ class _WaveClipper extends CustomClipper<Path> {
     for (int i = -2; i <= size.width.toInt() + 2; i++) {
       final waveHeight = (size.height / 20);
       final dy = math.sin((animationValue * 360 - i) % 360 * (math.pi / 180)) *
-          waveHeight +
+              waveHeight +
           (size.height - (size.height * value));
       waveList.add(Offset(i.toDouble(), dy));
     }
