@@ -59,12 +59,16 @@ public class FlutterAppUpgradePlugin : FlutterPlugin, MethodCallHandler, Activit
     if (call.method == "getAppInfo") {
       getAppInfo(mContext, result)
     } else if (call.method == "getApkDownloadPath") {
-      result.success(mContext.getExternalFilesDir("").absolutePath)
+      result.success(mContext.getExternalFilesDir("")?.absolutePath)
     } else if (call.method == "install") {
-      //安装app
-      val path = call.argument<String>("path")
-      path?.also {
-        startInstall(mContext, it)
+      if (BuildConfig.googlePlay) {
+        result.error("error", "googlePlay=true, The app installation function is disabled.", null)
+      } else {
+        // 安装app
+        val path = call.argument<String>("path")
+        path?.also {
+          startInstall(mContext, it)
+        }
       }
     } else if (call.method == "getInstallMarket") {
       var packageList = getInstallMarket(call.argument<List<String>>("packages"))
@@ -124,7 +128,7 @@ public class FlutterAppUpgradePlugin : FlutterPlugin, MethodCallHandler, Activit
       if (nameEmpty || classEmpty) {
         goToMarket.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
       } else {
-        goToMarket.setClassName(marketPackageName, marketClassName)
+        goToMarket.setClassName(marketPackageName!!, marketClassName!!)
       }
       context.startActivity(goToMarket)
     } catch (e: ActivityNotFoundException) {
@@ -157,11 +161,7 @@ public class FlutterAppUpgradePlugin : FlutterPlugin, MethodCallHandler, Activit
     val intent = Intent().setPackage(packageName)
     val infos = manager.queryIntentActivities(intent,
             PackageManager.GET_INTENT_FILTERS)
-    return if (infos == null || infos.size < 1) {
-      false
-    } else {
-      true
-    }
+    return infos.size >= 1
   }
 
   /**
