@@ -4,7 +4,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_app_upgrade/flutter_app_upgrade.dart';
 
 class FlutterUpgrade {
-  static const MethodChannel _channel = const MethodChannel('flutter_app_upgrade');
+  static const MethodChannel _channel =
+      const MethodChannel('flutter_app_upgrade');
 
   ///
   /// 获取app信息
@@ -43,7 +44,8 @@ class FlutterUpgrade {
   ///
   /// 获取android手机上安装的应用商店
   ///
-  static getInstallMarket({List<String> marketPackageNames}) async {
+  static Future<List<String>> getInstallMarket(
+      {List<String>? marketPackageNames}) async {
     List<String> packageNameList = AppMarket.buildInPackageNameList;
     if (marketPackageNames != null && marketPackageNames.length > 0) {
       packageNameList.addAll(marketPackageNames);
@@ -59,7 +61,7 @@ class FlutterUpgrade {
   ///
   /// 跳转到应用商店
   ///
-  static toMarket({AppMarketInfo appMarketInfo}) async {
+  static toMarket({AppMarketInfo? appMarketInfo}) async {
     var map = {
       'marketPackageName':
           appMarketInfo != null ? appMarketInfo.packageName : '',
